+++ src/main/java/com/yapp/artie/domain/user/application/port/in/GetUserDetailsQuery.java
@@ -1,0 +1,9 @@
+package com.yapp.artie.domain.user.application.port.in;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+public interface GetUserDetailsQuery extends UserDetailsService {
+
+  UserDetails loadUserByUsername(String uid);
+}
