+++ src/main/java/com/yapp/artie/domain/user/application/port/in/GetUserQuery.java
@@ -1,12 +1,8 @@
 package com.yapp.artie.domain.user.application.port.in;
 
 import com.yapp.artie.domain.user.domain.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 
-public interface GetUserQuery extends UserDetailsService {
+public interface GetUserQuery {
 
   User loadUserById(Long userId);
-
-  UserDetails loadUserByUsername(String uid);
 }
