+++ src/test/java/com/yapp/artie/domain/user/application/service/GetUserServiceTest.java
@@ -11,7 +11,6 @@ import com.yapp.artie.domain.user.domain.User;
 import com.yapp.artie.domain.user.exception.UserNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.springframework.security.core.userdetails.UserDetails;
 
 class GetUserServiceTest {
 
@@ -33,29 +32,8 @@ class GetUserServiceTest {
     assertThat(actual.getId()).isEqualTo(user.getId());
   }
 
-  @Test
-  void loadUserByUsername_사용자를_조회했을_때_반환되는_객체의_ID는_엔티티의_식별자와_같다() {
-    User user = defaultUser().build();
-    givenUserByReferenceWithUid(user);
-    UserDetails actual = getUserService.loadUserByUsername(user.getUid());
-    assertThat(actual.getUsername()).isEqualTo(user.getId().toString());
-  }
-
-  @Test
-  void loadUserByUsername_사용자를_조회했을_때_반환되는_객체의_PW는_엔티티의_UID와_같다() {
-    User user = defaultUser().build();
-    givenUserByReferenceWithUid(user);
-    UserDetails actual = getUserService.loadUserByUsername(user.getUid());
-    assertThat(actual.getPassword()).isEqualTo(user.getUid());
-  }
-
   private void givenUserByReference(User user) {
     given(loadUserPort.loadById(any()))
-        .willReturn(user);
-  }
-
-  private void givenUserByReferenceWithUid(User user) {
-    given(loadUserPort.loadByUid(any()))
         .willReturn(user);
   }
